playwright-python/requirements-dev.txt
# Test and code quality dependencies
pytest-cov>=3.0.0
pylint>=2.12.0
pylint-pytest>=1.1.2
pytest-xdist>=2.5.0
coverage>=6.3.2
pytest-html>=3.1.1
python-dotenv>=1.0.0
playwright>=1.41.0
